refactor: Writing of report
