+++ cognite/powerops/resync/__init__.py
@@ -1,4 +1,4 @@
 from ._main import MODEL_BY_NAME, apply, plan, AVAILABLE_MODELS, DEFAULT_MODELS
-from . import reports
 
-__all__ = ["apply", "plan", "MODEL_BY_NAME", "AVAILABLE_MODELS", "DEFAULT_MODELS", "reports"]
+
+__all__ = ["apply", "plan", "MODEL_BY_NAME", "AVAILABLE_MODELS", "DEFAULT_MODELS"]
